REFACTORED:Code As Per The Coding Principles

+++ src/main/java/com/bridgelabzs/statecensusanalyser/CSVStateCensus.java
@@ -7,15 +7,15 @@ public class CSVStateCensus {
     public String state;
 
     @CsvBindByName(column = "Population", required = true)
-    public int population;
+    public double population;
 
     @CsvBindByName(column = "AreaInSqKm", required = true)
-    public int area;
+    public double area;
 
     @CsvBindByName(column = "DensityPerSqKm", required = true)
-    public int density;
+    public double density;
 
-    public CSVStateCensus(String state, int population, int area, int density) {
+    public CSVStateCensus(String state, double population, double area, double density) {
         this.state = state;
         this.population = population;
         this.area = area;
